requirements.txt
Django
Djangorestframework
Djangorestframework-jwt
Djangorestframework-simplejwt
gunicorn==20.1.0